requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
asyncio==4.0.0
attrs==26.1.0
ccxt==4.5.45
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
coincurve==21.0.0
cryptography==46.0.6
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
numpy==2.4.4
pandas==3.0.1
propcache==0.4.1
pycares==5.0.1
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
setuptools==82.0.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
